update _put_back_hash to only call RPOPLPUSH to avoid keys getting lost
If the pod dies between LREM and LPUSH, the key vanishes.  This is no 
good. Using RPOPLPUSH to atomically move the item back to the queue.  If 
the queue is somehow empty...we assume the janitor got it. If we moved 
the wrong key, we log that and try to put another key from the queue 
back (HOW ARE THERE MORE THAN 1 key!?)